LPD-85801 SF

+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/internal/score/CTScoreCalculator.java
@@ -61,7 +61,7 @@ public class CTScoreCalculator {
 			score++;
 		}
 
-		long countMultiplier = _countTable(modelClassNameId) / _COUNT_DIVISOR;
+		int countMultiplier = _countTable(modelClassNameId) / _COUNT_DIVISOR;
 
 		if (countMultiplier == 0) {
 			countMultiplier = 1;
